+++ pyam/_aggregate.py
@@ -1,23 +1,33 @@
 import pandas as pd
 import numpy as np
 import logging
+from itertools import compress
 
+from pyam.index import replace_index_values
 from pyam.logging import adjust_log_level
-from pyam.utils import islistable, isstr, find_depth, reduce_hierarchy, KNOWN_FUNCS
+from pyam.utils import (
+    islistable,
+    isstr,
+    find_depth,
+    reduce_hierarchy,
+    KNOWN_FUNCS,
+    to_list,
+)
 
 logger = logging.getLogger(__name__)
 
 
 def _aggregate(df, variable, components=None, method=np.sum):
     """Internal implementation of the `aggregate` function"""
+
     # list of variables require default components (no manual list)
     if islistable(variable) and components is not None:
         raise ValueError(
-            "aggregating by list of variables cannot use " "custom components"
+            "Aggregating by list of variables does not support `components`!"
         )
 
     mapping = {}
-    msg = "cannot aggregate variable `{}` because it has no components"
+    msg = "Cannot aggregate variable '{}' because it has no components!"
     # if single variable
     if isstr(variable):
         # default components to all variables one level below `variable`
@@ -42,40 +52,29 @@ def _aggregate(df, variable, components=None, method=np.sum):
                 mapping[c] = v
 
     # rename all components to `variable` and aggregate
-    _df = df.data[df._apply_filters(variable=mapping.keys())].copy()
-    _df["variable"].replace(mapping, inplace=True)
+    _df = df._data[df._apply_filters(variable=mapping.keys())]
+    _df.index = replace_index_values(_df, "variable", mapping)
     return _group_and_agg(_df, [], method)
 
 
-def _aggregate_recursive(df, variable, method=np.sum):
+def _aggregate_recursive(df, variable):
     """Recursive aggregation along the variable tree"""
-    _df_aggregated = None
-    _df = df.copy()
 
-    # iterate over variables to find all subcategories to be aggregated
-    sub_variables = []
-    for d in reversed(range(1, max(find_depth(df.data.variable)) + 1)):
-        depth = find_depth(df.data.variable)
-        var_list = df.data.variable[[i == d for i in depth]].unique()
-        vars_up = pd.Series([reduce_hierarchy(i, -1) for i in var_list]).unique()
+    # downselect to components of `variable`, initialize list for aggregated (new) data
+    _df = df.filter(variable=f"{variable}|*")
+    data_list = []
 
-        if [i for i, entr in enumerate(vars_up) if entr.startswith(variable)]:
-            for v in vars_up:
-                sub_variables.append(v)
+    # iterate over variables (bottom-up) and aggregate all components up to `variable`
+    for d in reversed(range(find_depth(variable), max(find_depth(_df.variable)))):
+        components = compress(_df.variable, find_depth(_df.variable, level=d + 1))
+        var_list = set([reduce_hierarchy(v, -1) for v in components])
 
-    sub_variables = reversed(sorted(set(sub_variables)))
+        # a temporary dataframe allows to distinguish between full data and new data
+        temp_df = _df.aggregate(variable=var_list)
+        _df.append(temp_df, inplace=True)
+        data_list.append(temp_df._data)
 
-    # iterate over subcategories (bottom-up) and perform aggregation
-    for entry in sub_variables:
-        _df.aggregate(variable=entry, append=True)
-        _df_temp = _df.aggregate(variable=entry, append=False)
-
-        if _df_aggregated is None:
-            _df_aggregated = _df_temp.copy()
-        else:
-            _df_aggregated.append(_df_temp, inplace=True)
-
-    return _df_aggregated.data
+    return pd.concat(data_list)
 
 
 def _aggregate_region(
@@ -83,39 +82,36 @@ def _aggregate_region(
 ):
     """Internal implementation for aggregating data over subregions"""
     if not isstr(variable) and components is not False:
-        msg = "aggregating by list of variables with components " "is not supported"
-        raise ValueError(msg)
+        raise ValueError(
+            "Aggregating by list of variables with components is not supported!"
+        )
 
     if weight is not None and components is not False:
-        msg = "using weights and components in one operation not supported"
-        raise ValueError(msg)
+        raise ValueError("Using weights and components in one operation not supported!")
 
     # default subregions to all regions other than `region`
     subregions = subregions or df._all_other_regions(region, variable)
 
     if not len(subregions):
-        msg = (
-            "cannot aggregate variable `{}` to `{}` because it does not"
-            " exist in any subregion"
+        logger.info(
+            f"Cannot aggregate variable '{variable}' to '{region}' "
+            "because it does not exist in any subregion!"
         )
-        logger.info(msg.format(variable, region))
-
         return
 
     # compute aggregate over all subregions
     subregion_df = df.filter(region=subregions)
     rows = subregion_df._apply_filters(variable=variable)
     if weight is None:
-        col = "region"
-        _data = _group_and_agg(subregion_df.data[rows], col, method=method)
+        _data = _group_and_agg(subregion_df._data[rows], "region", method=method)
     else:
         weight_rows = subregion_df._apply_filters(variable=weight)
         _data = _agg_weight(
-            subregion_df.data[rows], subregion_df.data[weight_rows], method
+            subregion_df._data[rows], subregion_df._data[weight_rows], method
         )
 
     # if not `components=False`, add components at the `region` level
-    if components is not False:
+    if components:
         with adjust_log_level(logger):
             region_df = df.filter(region=region)
 
@@ -130,8 +126,9 @@ def _aggregate_region(
         if len(components):
             # rename all components to `variable` and aggregate
             rows = region_df._apply_filters(variable=components)
-            _df = region_df.data[rows].copy()
-            _df["variable"] = variable
+            _df = region_df._data[rows]
+            mapping = {c: variable for c in components}
+            _df.index = replace_index_values(_df.index, "variable", mapping)
             _data = _data.add(_group_and_agg(_df, "region"), fill_value=0)
 
     return _data
@@ -146,7 +143,7 @@ def _aggregate_time(df, variable, column, value, components, method=np.sum):
     # compute aggregate over time
     filter_args = dict(variable=variable)
     filter_args[column] = components
-    index = _list_diff(df.data.columns, [column, "value"])
+    index = df._data.index.names.difference([column])
 
     _data = pd.concat(
         [
@@ -166,40 +163,27 @@ def _aggregate_time(df, variable, column, value, components, method=np.sum):
 
 
 def _group_and_agg(df, by, method=np.sum):
-    """Groupby & aggregate `df` by column(s), return indexed `pd.Series`"""
-    by = [by] if isstr(by) else by
-    cols = [c for c in list(df.columns) if c not in ["value"] + by]
+    """Group-by & aggregate `pd.Series` by index names on `by`"""
+    cols = df.index.names.difference(to_list(by))
     # pick aggregator func (default: sum)
-    return df.groupby(cols)["value"].agg(_get_method_func(method))
+    return df.groupby(cols).agg(_get_method_func(method))
 
 
-def _agg_weight(df, weight, method):
-    """Aggregate `df` by regions with weights, return indexed `pd.Series`"""
+def _agg_weight(data, weight, method):
+    """Aggregate `data` by regions with weights, return indexed `pd.Series`"""
+
     # only summation allowed with weights
     if method not in ["sum", np.sum]:
-        raise ValueError("only method `np.sum` allowed for weighted average")
+        raise ValueError("Only method 'np.sum' allowed for weighted average!")
 
-    w_cols = _list_diff(df.columns, ["variable", "unit", "value"])
-    _weight = _get_value_col(weight, w_cols)
+    weight = weight.droplevel(["variable", "unit"])
 
-    if not _get_value_col(df, w_cols).index.equals(_weight.index):
-        raise ValueError("inconsistent index between variable and weight")
+    if not data.droplevel(["variable", "unit"]).index.equals(weight.index):
+        raise ValueError("Inconsistent index between variable and weight!")
 
-    _data = _get_value_col(df)
-    col1 = _list_diff(_data.index.names, ["region"])
-    col2 = _list_diff(w_cols, ["region"])
-    return (_data * _weight).groupby(col1).sum() / _weight.groupby(col2).sum()
-
-
-def _list_diff(lst, exclude):
-    """Return the list minus those elements in `exclude`"""
-    return [i for i in lst if i not in exclude]
-
-
-def _get_value_col(df, cols=None):
-    """Return the value column as `pd.Series sorted by index"""
-    cols = cols or [i for i in df.columns if i != "value"]
-    return df.set_index(cols)["value"].sort_index()
+    col1 = data.index.names.difference(["region"])
+    col2 = data.index.names.difference(["region", "variable", "unit"])
+    return (data * weight).groupby(col1).sum() / weight.groupby(col2).sum()
 
 
 def _get_method_func(method):
@@ -211,4 +195,4 @@ def _get_method_func(method):
         return KNOWN_FUNCS[method]
 
     # raise error if `method` is a string but not in dict of known methods
-    raise ValueError("method `{}` is not a known aggregator".format(method))
+    raise ValueError(f"'{method}' is not a known method!")