+++ RELEASE_NOTES.md
@@ -2,6 +2,12 @@
 
 ## API changes
 
+PR [#507](https://github.com/IAMconsortium/pyam/pull/507) harmonizes the behavior of
+the `aggregate()` and `aggregate_region()` methods when performing "empty" aggregation,
+i.e., no components exist to perform the requested aggregation.
+In the new implementation, an empty IamDataFrame is returned if `append=False`
+(instead of None).
+
 PR [#488](https://github.com/IAMconsortium/pyam/pull/488) changes the default
 behavior when initializing an IamDataFrame from xlsx: now, all sheets names
 starting with `data` will be parsed for timeseries data.
@@ -10,6 +16,7 @@ starting with `data` will be parsed for timeseries data.
 
 - [#510](https://github.com/IAMconsortium/pyam/pull/510) Improve performance of `pyam.concat()`
 - [#508](https://github.com/IAMconsortium/pyam/pull/508) Bugfix for non-empty but invisible header and no rows in 'meta' sheet
+- [#507](https://github.com/IAMconsortium/pyam/pull/507) Refactor aggregation-functions to increase performance
 - [#502](https://github.com/IAMconsortium/pyam/pull/502) Switch to Black code style
 - [#499](https://github.com/IAMconsortium/pyam/pull/499) Implement `order` feature in line plot 
 - [#497](https://github.com/IAMconsortium/pyam/pull/497) Add a module for reading data from the UNFCCC Data Inventory